+++ src/SO_ARM100_kinematics/include/HybridSolver/WristModel.hpp
@@ -18,7 +18,7 @@ struct WristModel
 {
 	WristType type { WristType::None };
 	int active_joint_start{ 0 };
-	int active_joint_count{ 0 };
+	size_t active_joint_count{ 0 };
 	Vec3d center_at_home{ Vec3d::Zero() };
 	Mat4d tcp_in_wrist_at_home { Mat4d::Identity() };
 	Mat4d tcp_in_wrist_at_home_inv { Mat4d::Identity() };
